refactor: Moved isAllDigit into a generaul utils header
