+++ inc/config/internal/utils.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <string>
-#include "utils/utils.hpp"
 
 namespace utils {
 
@@ -9,7 +8,6 @@ bool isValidPort(int port);
 bool isValidPort(std::string const &);
 bool isValidIPv4(char const *);
 bool isValidIPv4(std::string const &);
-bool isAllDigit(std::string const &s);
 
 struct IpInfo {
     std::string ip;
