+++ src/http/HttpResponse.cpp
@@ -1,8 +1,8 @@
-#include "http/HttpStatus.hpp"
 #include "http/HttpResponse.hpp"
-#include "config/internal/utils.hpp"
-#include <string>
+#include "http/HttpStatus.hpp"
+#include "utils/utils.hpp"
 #include <sstream>
+#include <string>
 
 namespace http {
 
@@ -37,12 +37,10 @@ inline const char *getResponsePhrase(Status status) {
 
 } // namespace
 
-HttpResponse::HttpResponse() : statusCode_(INTERNAL_SERVER_ERROR), bodyType_(BODY_NONE) {
-}
+HttpResponse::HttpResponse() : statusCode_(INTERNAL_SERVER_ERROR), bodyType_(BODY_NONE) {}
 
 HttpResponse::HttpResponse(Status code, std::string const &httpVersion)
-    : httpVersion_(httpVersion), statusCode_(code), bodyType_(BODY_NONE) {
-}
+    : httpVersion_(httpVersion), statusCode_(code), bodyType_(BODY_NONE) {}
 
 HttpResponse::HttpResponse(HttpResponse const &rhs)
     : httpVersion_(rhs.httpVersion_),
@@ -78,13 +76,9 @@ HttpResponse const &HttpResponse::operator=(HttpResponse const &rhs) {
     return *this;
 }
 
-HttpResponse::~HttpResponse() {
-    cleanupBody();
-}
+HttpResponse::~HttpResponse() { cleanupBody(); }
 
-void HttpResponse::setStatus(Status s) {
-    statusCode_ = s;
-}
+void HttpResponse::setStatus(Status s) { statusCode_ = s; }
 
 void HttpResponse::setNoBody() {
     cleanupBody();
@@ -123,29 +117,17 @@ void HttpResponse::setBodyFromCgi(int pipe_fd) {
     cgiBody.pipe_fd = pipe_fd;
 }
 
-Status HttpResponse::getStatus() const {
-    return statusCode_;
-}
+Status HttpResponse::getStatus() const { return statusCode_; }
 
-char const *HttpResponse::getResponsePhrase() const {
-    return http::getResponsePhrase(statusCode_);
-}
+char const *HttpResponse::getResponsePhrase() const { return http::getResponsePhrase(statusCode_); }
 
-HeaderMap &HttpResponse::getHeaders() {
-    return headers_;
-}
+HeaderMap &HttpResponse::getHeaders() { return headers_; }
 
-HeaderMap const &HttpResponse::getHeaders() const {
-    return headers_;
-}
+HeaderMap const &HttpResponse::getHeaders() const { return headers_; }
 
-std::string const &HttpResponse::getVersion() const {
-    return httpVersion_;
-}
+std::string const &HttpResponse::getVersion() const { return httpVersion_; }
 
-BodySourceType HttpResponse::getBodyType() const {
-    return bodyType_;
-}
+BodySourceType HttpResponse::getBodyType() const { return bodyType_; }
 
 void HttpResponse::cleanupBody() {
     if (bodyType_ == BODY_IN_MEMORY && inMemoryBody.data != NULL) {
